tmp commit

+++ dao/scheme/role.go
@@ -12,10 +12,10 @@ type Role struct {
 	Name        string     `gorm:"type:varchar(64);uniqueIndex" json:"name"`
 	Description string     `gorm:"type:text" json:"description"`
 	CreatedAt   time.Time  `gorm:"autoCreateTime" json:"created_at"`
-	CreatedBy   *int64     `json:"-"`
+	CreatedBy   *uint      `json:"-"`
 	CreatedUser *User      `gorm:"foreignKey:CreatedBy"`
 	UpdatedAt   *time.Time `gorm:"autoUpdateTime" json:"updated_at"`
-	UpdatedBy   *int64     `json:"-"`
+	UpdatedBy   *uint      `json:"-"`
 	UpdatedUser *User      `gorm:"foreignKey:UpdatedBy"`
 	Users       []*User    `gorm:"many2many:roles_users;"`
 }
